+++ Src/Library/Src/Controller/model_excitation.c
@@ -1,8 +1,8 @@
 #include "Controller/model_excitation.h"
 
-float excitation_signal           = 0;
-float excitation_signal_magnitude = 250;
-float excitation_signal_gain      = 1;
+double excitation_signal           = 0;
+double excitation_signal_magnitude = 250;
+double excitation_signal_gain      = 1;
 
 void init_model_excitation(void)
 {
